requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow[and-cuda]==2.13.0
tensorflow-gpu==2.13.0
keras==2.13.0
torch==2.0.1
cupy-cuda11x==12.2.0
cuda-python==12.2.0
transformers==4.31.0
xgboost==2.0.0
lightgbm==4.0.0
cuml==23.10.0
ta==0.10.2
ta-lib==0.4.28
yfinance==0.2.28
alpha-vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pytesseract==0.3.10
opencv-python==4.8.0.76
pillow==10.0.0
mss==9.0.1
pyautogui==0.9.53
newsapi==0.1.1
textblob==0.17.1
nltk==3.8.1
scipy==1.11.2
statsmodels==0.14.0
dash==2.14.1
plotly==5.16.1
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
joblib==1.3.1
pydantic==2.2.1
sqlalchemy==2.0.20
pymongo==4.4.1
redis==5.0.0
logging-loki==0.3.2
pyyaml==6.0.1
tqdm==4.66.1
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.4.1
